// app/actions/schemas.ts
// app/schemas.ts

import { z } from 'zod'

export const userSchema = z.object({
    id: z.string(),
    name: z.string().min(2, { message: "Name must be at least 2 characters." }),
    email: z.string().email({ message: "Invalid email address." }),
    phoneNumber: z.string().regex(/^04\d{8}$/, { message: "Phone number must be a valid Australian mobile number" }),
    createdAt: z.date(),
    updatedAt: z.date()
})

export type User = z.infer<typeof userSchema>

export const userFormSchema = userSchema.omit({ 
    id: true, 
    createdAt: true, 
    updatedAt: true 
})
export type UserFormData = z.infer<typeof userFormSchema>